add update employer

+++ src/services/user.js
@@ -35,26 +35,22 @@ export const getApplicantById = async (id) => {
   }
 };
 
-export const updateUser = async (body, fileData) =>
+export const updateAvatar = async (body, fileData) =>
   new Promise(async (resolve, reject) => {
     try {
       // Find the existing resume record using findByPk
-      const user_id = body?.user_id;
+      const user_id = body.user_id;
       const user = await db.User.findOne({ where: { id: user_id } });
-
       if (fileData) {
         const filename = user.file_name;
         if (filename) {
-          await cloudinary.uploader.destroy(user.file_name);
+          await cloudinary.uploader.destroy(filename);
         }
         body.avatar = fileData?.path;
         body.file_name = fileData?.filename;
       }
       const response = await db.User.update(body, { where: { id: user_id } });
-      resolve({
-        err: 0,
-        mes: "User updated successfully",
-      });
+      resolve(response);
       if (fileData && response[0] === 0) {
         cloudinary.uploader.destroy(fileData?.filename);
       }
@@ -65,6 +61,17 @@ export const updateUser = async (body, fileData) =>
     }
   });
 
+export const updateUser = async (body, id) => {
+  try {
+    const response = await db.User.update(body, {
+      where: { id: id },
+    });
+    return response;
+  } catch (error) {
+    return error;
+  }
+};
+
 export const deleteUser = async (id) =>
   new Promise(async (resolve, reject) => {
     try {
@@ -73,9 +80,14 @@ export const deleteUser = async (id) =>
       const user_role = user.role_code;
       if (user_role === "R3")
         await db.Applicant.destroy({ where: { user_id: id } });
-      if (user_role === "R2")
-        await db.Employer.destroy({ where: { user_id: id } });
-
+      if (user_role === "R2") {
+        const employer = await db.Employer.findOne({ where: { user_id: id } });
+        if (employer) {
+          if (employer.file_name)
+            await cloudinary.uploader.destroy(employer.file_name);
+          await db.Employer.destroy({ where: { user_id: id } });
+        }
+      }
       const response = await db.User.destroy({
         where: { id: id },
       });
